Naive predictor: preprocessing, splitting and metr
preprocessing: preprocessing of data happens before data splitting and
is optional. The methods have been added to utils.pz and are called in
model.py

Dataspliting: was aded to utils as we want to split the data once such
thats its always the same for all models

calc metric: done calculating metrics in the LDO and LCO settings. LPO
has to be done now (getting LPO data is already done)