Oversampling for classifiers implemented and confusion metrices updated
added the functionality of oversampling the training X and y data in the
classifier case, as here there is usually an imbalance between sensitive
and resistant drugs (usually only 10% of the lables are sensitive and
the rest is resistant). Different oversampling strategies can be used by
the user (random oversampling, SMOTE, etc.).

Added accuracy, balanced accuracy metrices for the confusion matrix
calculated in order to give a better sense on dataset in balances.

Added more detailed description to base model methods and functions.