Changed score calculation and inserted SMOTEN
THe performance metrices calculated for the logclf have now been changed
to confusion matrix values, F1-score, AUC and MCC. This happens in the
evaluate() method of the linearClassifier class. The confusionmatrix
function in analysis.py has been removed and replaced by the
score_clustering function which performs the clustering of the
metrics_df for the classifier as well as the regressor.

Additionally changed for SMOTEN to be used in the case for MF in the LDO
case as here we have nominal input features and the SMOTEN function has
to be used therefore.