Created SVC model
created the support vector classifier and its coresponding model.py and
testing.py files. Had to change the parsing and testing.py since the SVC
can use different kernels and the user can specify which one in the toml
file and this is stored in parameter self.kernel. Since rbf and poly kernels
 dont output weights, the function
calcualting how many coefficients are == 0 doesnt work here, so had to
change the base_analysis function to not include that but rather be a
separate function called my appropraite models.

Also created a function for plotting the boundary decision and maximum
margine. However, this only works for 2D data, so only data which has
been previously transformed by PCA into an X with two fatures. Plotting
higher diemnsional features isnt possobile beyond 3D anyway and the
sklearn function only takes in 2D data.

mkdir now accepts a full path, not only a local one, to store the output
in.