Changed performance metric clustering:
For regression, only heatmaps are now created as (plotting clustergrams
into subplots in plotly not really possible). Two different heatmaps are
subploted: one for the errors (mse,rmse) and one for the corrleation
scores (scc and pcc) each with ther own scale.

For classifiers a heatmap (sorted by pcc values) AND a clustergram are
outputted since here all the metrices have the same scale (from 0 to 1).

Minor changes regarding function description and parsing in other files.